user_service/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.6

social-auth-app-django>=4.0  # For OAuth2 and social login
djangorestframework-simplejwt
